refactor/#203: 불필요한 파일 제거 및 폴더링
